refactor(sync): extract historical state indexing into dedicated `IndexHistory` stage
The `Blocks` stage previously conflated two concerns: storing downloaded
blocks and building historical state indices. This separates them into
distinct pipeline stages following the Erigon/Reth pattern of dedicated
index stages.

Split `insert_block_with_states_and_receipts` into `insert_block_data`
(block storage only) and `insert_state_history` (historical indices).
The `BlockWriter` trait method delegates to both, preserving behavior
for the sequencing path and `ForkedProvider`.

Add `insert_state_history_bulk` for first-sync optimization that
accumulates changes in memory and writes once per key, avoiding the
read-modify-write overhead when history tables are known to be empty.

Move all pruning logic from `Blocks` to `IndexHistory` since the new
stage owns the tables being pruned. Register `IndexHistory` after
`Classes` and before `StateTrie` in the pipeline.
