+++ crates/storage/provider/provider/src/providers/db/mod.rs
@@ -519,8 +519,13 @@ impl<Tx: DbTx> BlockEnvProvider for DbProvider<Tx> {
     }
 }
 
-impl<Tx: DbTxMut> BlockWriter for DbProvider<Tx> {
-    fn insert_block_with_states_and_receipts(
+impl<Tx: DbTxMut> DbProvider<Tx> {
+    /// Stores block data without building historical state indices.
+    ///
+    /// This stores: headers, hashes, body indices, `BlockStateUpdates`, txs, receipts, traces,
+    /// class artifacts, compiled class hashes, class declarations, deprecated declarations,
+    /// migrated compiled class hashes.
+    pub fn insert_block_data(
         &self,
         block: SealedBlockWithStatus,
         states: StateUpdatesWithClasses,
@@ -602,14 +607,149 @@ impl<Tx: DbTxMut> BlockWriter for DbProvider<Tx> {
             self.0.put::<tables::MigratedCompiledClassHashes>(block_number, entry)?;
         }
 
+        Ok(())
+    }
+
+    /// Builds historical state indices for a range of blocks in bulk.
+    ///
+    /// This is an optimized path for first sync (when the history tables are empty). Instead of
+    /// doing read-modify-write per block, it accumulates all changes in memory and writes each
+    /// key once at the end. The DupSort history tables (`StorageChangeHistory`,
+    /// `NonceChangeHistory`, `ClassChangeHistory`) are written per-block since blocks are
+    /// processed in order and the keys are monotonically increasing.
+    ///
+    /// Updates the same tables as [`insert_state_history`](Self::insert_state_history).
+    pub fn insert_state_history_bulk(
+        &self,
+        blocks: impl IntoIterator<Item = (BlockNumber, StateUpdates)>,
+    ) -> ProviderResult<()> {
+        // Accumulated state: written once per key at the end.
+        let mut storage_change_sets: BTreeMap<ContractStorageKey, BlockChangeList> =
+            BTreeMap::new();
+        let mut contract_info_change_sets: BTreeMap<ContractAddress, ContractInfoChangeList> =
+            BTreeMap::new();
+        // Track latest storage value per (addr, key) — only the final value is written.
+        let mut latest_storage: BTreeMap<(ContractAddress, katana_primitives::Felt), StorageEntry> =
+            BTreeMap::new();
+        // Track latest contract info per address — only the final state is written.
+        let mut latest_contract_info: BTreeMap<ContractAddress, GenericContractInfo> =
+            BTreeMap::new();
+
+        for (block_number, state_updates) in blocks {
+            // -- storage changes --
+            for (addr, entries) in &state_updates.storage_updates {
+                for (key, value) in entries {
+                    let entry = StorageEntry { key: *key, value: *value };
+
+                    // Accumulate change set
+                    let changeset_key = ContractStorageKey { contract_address: *addr, key: *key };
+                    storage_change_sets
+                        .entry(changeset_key.clone())
+                        .or_default()
+                        .insert(block_number);
+
+                    // Track latest value
+                    latest_storage.insert((*addr, *key), entry);
+
+                    // Write per-block history entry (block-keyed DupSort, sequential)
+                    self.0.put::<tables::StorageChangeHistory>(
+                        block_number,
+                        ContractStorageEntry { key: changeset_key, value: *value },
+                    )?;
+                }
+            }
+
+            // -- deployed contracts --
+            for (addr, class_hash) in &state_updates.deployed_contracts {
+                let info = latest_contract_info.entry(*addr).or_default();
+                info.class_hash = *class_hash;
+
+                contract_info_change_sets
+                    .entry(*addr)
+                    .or_default()
+                    .class_change_list
+                    .insert(block_number);
+
+                let class_change_key = ContractClassChange::deployed(*addr, *class_hash);
+                self.0.put::<tables::ClassChangeHistory>(block_number, class_change_key)?;
+            }
+
+            // -- replaced classes --
+            for (addr, new_class_hash) in &state_updates.replaced_classes {
+                let info = latest_contract_info.entry(*addr).or_default();
+                info.class_hash = *new_class_hash;
+
+                contract_info_change_sets
+                    .entry(*addr)
+                    .or_default()
+                    .class_change_list
+                    .insert(block_number);
+
+                let class_change_key = ContractClassChange::replaced(*addr, *new_class_hash);
+                self.0.put::<tables::ClassChangeHistory>(block_number, class_change_key)?;
+            }
+
+            // -- nonce updates --
+            for (addr, nonce) in &state_updates.nonce_updates {
+                let info = latest_contract_info.entry(*addr).or_default();
+                info.nonce = *nonce;
+
+                contract_info_change_sets
+                    .entry(*addr)
+                    .or_default()
+                    .nonce_change_list
+                    .insert(block_number);
+
+                let nonce_change_key =
+                    ContractNonceChange { contract_address: *addr, nonce: *nonce };
+                self.0.put::<tables::NonceChangeHistory>(block_number, nonce_change_key)?;
+            }
+        }
+
+        // Flush accumulated storage change sets (one write per key).
+        for (key, block_list) in storage_change_sets {
+            self.0.put::<tables::StorageChangeSet>(key, block_list)?;
+        }
+
+        // Flush latest storage values (one write per (addr, key)).
+        {
+            let mut storage_cursor = self.0.cursor_dup_mut::<tables::ContractStorage>()?;
+            for ((addr, _), entry) in &latest_storage {
+                storage_cursor.upsert(*addr, *entry)?;
+            }
+        }
+
+        // Flush accumulated contract info change sets (one write per address).
+        for (addr, change_set) in contract_info_change_sets {
+            self.0.put::<tables::ContractInfoChangeSet>(addr, change_set)?;
+        }
+
+        // Flush latest contract info (one write per address).
+        for (addr, info) in latest_contract_info {
+            self.0.put::<tables::ContractInfo>(addr, info)?;
+        }
+
+        Ok(())
+    }
+
+    /// Builds historical state indices for a single block from its state updates.
+    ///
+    /// This updates: `ContractStorage`, `StorageChangeSet`, `StorageChangeHistory`,
+    /// `ContractInfo`, `ContractInfoChangeSet`, `ClassChangeHistory`, `NonceChangeHistory`.
+    pub fn insert_state_history(
+        &self,
+        block_number: BlockNumber,
+        state_updates: &StateUpdates,
+    ) -> ProviderResult<()> {
         // insert storage changes
         {
             let mut storage_cursor = self.0.cursor_dup_mut::<tables::ContractStorage>()?;
-            for (addr, entries) in state_updates.storage_updates {
-                let entries = entries.into_iter().map(|(key, value)| StorageEntry { key, value });
+            for (addr, entries) in &state_updates.storage_updates {
+                let entries =
+                    entries.iter().map(|(key, value)| StorageEntry { key: *key, value: *value });
 
                 for entry in entries {
-                    match storage_cursor.seek_by_key_subkey(addr, entry.key)? {
+                    match storage_cursor.seek_by_key_subkey(*addr, entry.key)? {
                         Some(current) if current.key == entry.key => {
                             storage_cursor.delete_current()?;
                         }
@@ -619,7 +759,7 @@ impl<Tx: DbTxMut> BlockWriter for DbProvider<Tx> {
 
                     // update block list in the change set
                     let changeset_key =
-                        ContractStorageKey { contract_address: addr, key: entry.key };
+                        ContractStorageKey { contract_address: *addr, key: entry.key };
                     let list = self.0.get::<tables::StorageChangeSet>(changeset_key.clone())?;
 
                     let updated_list = match list {
@@ -633,10 +773,10 @@ impl<Tx: DbTxMut> BlockWriter for DbProvider<Tx> {
                     };
 
                     self.0.put::<tables::StorageChangeSet>(changeset_key, updated_list)?;
-                    storage_cursor.upsert(addr, entry)?;
+                    storage_cursor.upsert(*addr, entry)?;
 
                     let storage_change_sharded_key =
-                        ContractStorageKey { contract_address: addr, key: entry.key };
+                        ContractStorageKey { contract_address: *addr, key: entry.key };
 
                     self.0.put::<tables::StorageChangeHistory>(
                         block_number,
@@ -651,15 +791,15 @@ impl<Tx: DbTxMut> BlockWriter for DbProvider<Tx> {
 
         // update contract info
 
-        for (addr, class_hash) in state_updates.deployed_contracts {
-            let value = if let Some(info) = self.0.get::<tables::ContractInfo>(addr)? {
-                GenericContractInfo { class_hash, ..info }
+        for (addr, class_hash) in &state_updates.deployed_contracts {
+            let value = if let Some(info) = self.0.get::<tables::ContractInfo>(*addr)? {
+                GenericContractInfo { class_hash: *class_hash, ..info }
             } else {
-                GenericContractInfo { class_hash, ..Default::default() }
+                GenericContractInfo { class_hash: *class_hash, ..Default::default() }
             };
 
             let new_change_set =
-                if let Some(mut change_set) = self.0.get::<tables::ContractInfoChangeSet>(addr)? {
+                if let Some(mut change_set) = self.0.get::<tables::ContractInfoChangeSet>(*addr)? {
                     change_set.class_change_list.insert(block_number);
                     change_set
                 } else {
@@ -669,22 +809,22 @@ impl<Tx: DbTxMut> BlockWriter for DbProvider<Tx> {
                     }
                 };
 
-            self.0.put::<tables::ContractInfo>(addr, value)?;
+            self.0.put::<tables::ContractInfo>(*addr, value)?;
 
-            let class_change_key = ContractClassChange::deployed(addr, class_hash);
+            let class_change_key = ContractClassChange::deployed(*addr, *class_hash);
             self.0.put::<tables::ClassChangeHistory>(block_number, class_change_key)?;
-            self.0.put::<tables::ContractInfoChangeSet>(addr, new_change_set)?;
+            self.0.put::<tables::ContractInfoChangeSet>(*addr, new_change_set)?;
         }
 
-        for (addr, new_class_hash) in state_updates.replaced_classes {
-            let info = if let Some(info) = self.0.get::<tables::ContractInfo>(addr)? {
-                GenericContractInfo { class_hash: new_class_hash, ..info }
+        for (addr, new_class_hash) in &state_updates.replaced_classes {
+            let info = if let Some(info) = self.0.get::<tables::ContractInfo>(*addr)? {
+                GenericContractInfo { class_hash: *new_class_hash, ..info }
             } else {
-                GenericContractInfo { class_hash: new_class_hash, ..Default::default() }
+                GenericContractInfo { class_hash: *new_class_hash, ..Default::default() }
             };
 
             let new_change_set =
-                if let Some(mut change_set) = self.0.get::<tables::ContractInfoChangeSet>(addr)? {
+                if let Some(mut change_set) = self.0.get::<tables::ContractInfoChangeSet>(*addr)? {
                     change_set.class_change_list.insert(block_number);
                     change_set
                 } else {
@@ -694,22 +834,22 @@ impl<Tx: DbTxMut> BlockWriter for DbProvider<Tx> {
                     }
                 };
 
-            self.0.put::<tables::ContractInfo>(addr, info)?;
+            self.0.put::<tables::ContractInfo>(*addr, info)?;
 
-            let class_change_key = ContractClassChange::replaced(addr, new_class_hash);
+            let class_change_key = ContractClassChange::replaced(*addr, *new_class_hash);
             self.0.put::<tables::ClassChangeHistory>(block_number, class_change_key)?;
-            self.0.put::<tables::ContractInfoChangeSet>(addr, new_change_set)?;
+            self.0.put::<tables::ContractInfoChangeSet>(*addr, new_change_set)?;
         }
 
-        for (addr, nonce) in state_updates.nonce_updates {
-            let value = if let Some(info) = self.0.get::<tables::ContractInfo>(addr)? {
-                GenericContractInfo { nonce, ..info }
+        for (addr, nonce) in &state_updates.nonce_updates {
+            let value = if let Some(info) = self.0.get::<tables::ContractInfo>(*addr)? {
+                GenericContractInfo { nonce: *nonce, ..info }
             } else {
-                GenericContractInfo { nonce, ..Default::default() }
+                GenericContractInfo { nonce: *nonce, ..Default::default() }
             };
 
             let new_change_set =
-                if let Some(mut change_set) = self.0.get::<tables::ContractInfoChangeSet>(addr)? {
+                if let Some(mut change_set) = self.0.get::<tables::ContractInfoChangeSet>(*addr)? {
                     change_set.nonce_change_list.insert(block_number);
                     change_set
                 } else {
@@ -719,13 +859,29 @@ impl<Tx: DbTxMut> BlockWriter for DbProvider<Tx> {
                     }
                 };
 
-            self.0.put::<tables::ContractInfo>(addr, value)?;
+            self.0.put::<tables::ContractInfo>(*addr, value)?;
 
-            let nonce_change_key = ContractNonceChange { contract_address: addr, nonce };
+            let nonce_change_key = ContractNonceChange { contract_address: *addr, nonce: *nonce };
             self.0.put::<tables::NonceChangeHistory>(block_number, nonce_change_key)?;
-            self.0.put::<tables::ContractInfoChangeSet>(addr, new_change_set)?;
+            self.0.put::<tables::ContractInfoChangeSet>(*addr, new_change_set)?;
         }
 
+        Ok(())
+    }
+}
+
+impl<Tx: DbTxMut> BlockWriter for DbProvider<Tx> {
+    fn insert_block_with_states_and_receipts(
+        &self,
+        block: SealedBlockWithStatus,
+        states: StateUpdatesWithClasses,
+        receipts: Vec<Receipt>,
+        executions: Vec<TypedTransactionExecutionInfo>,
+    ) -> ProviderResult<()> {
+        let block_number = block.block.header.number;
+        let state_updates = states.state_updates.clone();
+        self.insert_block_data(block, states, receipts, executions)?;
+        self.insert_state_history(block_number, &state_updates)?;
         Ok(())
     }
 }
